requirements.txt
Flask>=2.0
flask-cors>=3.0
python-dotenv>=0.19
openai>=1.40.0
xai-sdk>=0.3.0